backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI/ML and RAG
openai==1.3.7
pinecone-client==2.2.4
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pytesseract==0.3.10
pillow==10.1.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
celery==5.3.4
boto3==1.34.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring and observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
prometheus-client==0.19.0

# Security
cryptography==41.0.8